[Refactor] #1642 Moved quaff-execution.* from object-use/ to object-use/quaff/
