+++ Hengband/Hengband/Hengband.vcxproj
@@ -499,7 +499,7 @@
     <ClCompile Include="..\..\src\object-hook\hook-quest.cpp" />
     <ClCompile Include="..\..\src\object-hook\hook-weapon.cpp" />
     <ClCompile Include="..\..\src\object-hook\hook-armor.cpp" />
-    <ClCompile Include="..\..\src\object-use\quaff-execution.cpp" />
+    <ClCompile Include="..\..\src\object-use\quaff\quaff-execution.cpp" />
     <ClCompile Include="..\..\src\load\angband-version-comparer.cpp" />
     <ClCompile Include="..\..\src\load\birth-loader.cpp" />
     <ClCompile Include="..\..\src\load\dummy-loader.cpp" />
@@ -1267,7 +1267,7 @@
     <ClInclude Include="..\..\src\object-hook\hook-perception.h" />
     <ClInclude Include="..\..\src\object-hook\hook-quest.h" />
     <ClInclude Include="..\..\src\object-hook\hook-weapon.h" />
-    <ClInclude Include="..\..\src\object-use\quaff-execution.h" />
+    <ClInclude Include="..\..\src\object-use\quaff\quaff-execution.h" />
     <ClInclude Include="..\..\src\player\attack-defense-types.h" />
     <ClInclude Include="..\..\src\core\player-update-types.h" />
     <ClInclude Include="..\..\src\load\angband-version-comparer.h" />
